Chapter 6: Added controller documentation

+++ src/orders/controller/orders.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  NotFoundException,
   Param,
   Post,
   Put,
@@ -12,6 +13,7 @@ import { OrdersService } from '../service/orders.service';
 import Order from '../domain/order.domain';
 import { CustomersService } from 'src/customers/service/customers.service';
 import OrderDTO from '../dto/order.dto';
+import { ApiResponse } from '@nestjs/swagger';
 
 @Controller('orders')
 export class OrdersController {
@@ -21,16 +23,27 @@ export class OrdersController {
     private customersService: CustomersService,
   ) {}
 
+  @ApiResponse({ status: 200, description: 'Returns the orders' })
   @Get()
   async getOrders(): Promise<Order[]> {
     return await this.ordersService.getOrders();
   }
 
+  @ApiResponse({ status: 200, description: 'Returns the order' })
+  @ApiResponse({ status: 404, description: "The order can't be found" })
   @Get(':id')
   async getOrderById(@Param() { id }: { id: string }): Promise<Order | null> {
-    return await this.ordersService.getOrderById(id);
+    const order = await this.ordersService.getOrderById(id);
+    if (!order) {
+      throw new NotFoundException();
+    }
+    return order;
   }
 
+  @ApiResponse({
+    status: 201,
+    description: 'The order was created successfully',
+  })
   @Post()
   async createOrder(@Body() orderDTO: OrderDTO): Promise<Order> {
     const customer = await this.customersService.getCustomerById(
@@ -46,11 +59,20 @@ export class OrdersController {
     return order;
   }
 
+  @ApiResponse({
+    status: 200,
+    description: 'The order was updated successfully',
+  })
+  @ApiResponse({ status: 404, description: "The order can't be found" })
   @Put(':id')
   async updateOrder(
     @Param() { id }: { id: string },
     @Body() orderDTO: OrderDTO,
   ): Promise<Order> {
+    const checkOrder = this.ordersService.getOrderById(id);
+    if (!checkOrder) {
+      throw new NotFoundException();
+    }
     const customer = await this.customersService.getCustomerById(
       orderDTO.customer,
     );
@@ -62,8 +84,17 @@ export class OrdersController {
     return await this.ordersService.updateOrder(order);
   }
 
+  @ApiResponse({
+    status: 200,
+    description: 'The order was deleted successfully',
+  })
+  @ApiResponse({ status: 404, description: "The order can't be found" })
   @Delete(':id')
   async removeOrder(@Param() { id }: { id: string }) {
+    const checkOrder = this.ordersService.getOrderById(id);
+    if (!checkOrder) {
+      throw new NotFoundException();
+    }
     await this.ordersService.removeOrder(id);
   }
 }
