+++ src/products/controller/productCategories.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  NotFoundException,
   Param,
   Post,
   Put,
@@ -12,6 +13,7 @@ import { ProductCategoriesService } from '../service/productCategories.service';
 import ProductCategory from '../domain/productCategory.domain';
 import ProductCategoryDTO from '../dto/productCatergory.dto';
 import { ProductsService } from '../service/products.service';
+import { ApiResponse } from '@nestjs/swagger';
 
 @Controller('productCategories')
 export class ProductCategoriesController {
@@ -21,18 +23,33 @@ export class ProductCategoriesController {
     private productsService: ProductsService,
   ) {}
 
+  @ApiResponse({ status: 200, description: 'Returns the product categories' })
   @Get()
   async getProducts(): Promise<ProductCategory[]> {
     return await this.productCategoriesService.getProductCategories();
   }
 
+  @ApiResponse({ status: 200, description: 'Returns the product category' })
+  @ApiResponse({
+    status: 404,
+    description: "The product category can't be found",
+  })
   @Get(':id')
   async getProductById(
     @Param() { id }: { id: string },
   ): Promise<ProductCategory | null> {
-    return await this.productCategoriesService.getProductCategoryById(id);
+    const productCategory =
+      await this.productCategoriesService.getProductCategoryById(id);
+    if (!productCategory) {
+      throw new NotFoundException();
+    }
+    return productCategory;
   }
 
+  @ApiResponse({
+    status: 201,
+    description: 'The product category was created successfully',
+  })
   @Post()
   async createProduct(
     @Body() productCategoryDTO: ProductCategoryDTO,
@@ -42,11 +59,24 @@ export class ProductCategoriesController {
     );
   }
 
+  @ApiResponse({
+    status: 200,
+    description: 'The product category was updated successfully',
+  })
+  @ApiResponse({
+    status: 404,
+    description: "The product category can't be found",
+  })
   @Put(':id')
   async updateProduct(
     @Param() { id }: { id: string },
     @Body() productCategoryDTO: ProductCategoryDTO,
   ): Promise<ProductCategory> {
+    const checkProductCategory =
+      this.productCategoriesService.getProductCategoryById(id);
+    if (!checkProductCategory) {
+      throw new NotFoundException();
+    }
     const productCategory =
       this.productCategoriesMapper.dtoToProductCategory(productCategoryDTO);
     productCategory.id = id;
@@ -55,8 +85,21 @@ export class ProductCategoriesController {
     );
   }
 
+  @ApiResponse({
+    status: 200,
+    description: 'The product category was deleted successfully',
+  })
+  @ApiResponse({
+    status: 404,
+    description: "The product category can't be found",
+  })
   @Delete(':id')
   async removeProduct(@Param() { id }: { id: string }) {
+    const checkProductCategory =
+      this.productCategoriesService.getProductCategoryById(id);
+    if (!checkProductCategory) {
+      throw new NotFoundException();
+    }
     await this.productCategoriesService.removeProductCategory(id);
   }
 }
